requirements.txt
sounddevice==0.4.6
numpy==1.26.4
scipy==1.13.0
keyboard==0.13.5
pyperclip==1.8.2
pyautogui==0.9.54
openai==1.30.0
google-generativeai==0.7.0
groq==0.9.0
python-dotenv==1.0.1
